# api/bundled/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "description": "Egolia API",
    "version": "0.0.0",
    "title": "Egolia API",
    "license": {
      "name": "MIT",
      "url": "https://github.com/egolia-uit/egolia/blob/main/LICENSE"
    },
    "contact": {
      "name": "Egolia Support",
      "email": "egolia@googlegroups.com"
    },
    "x-apiName": "api"
  },
  "servers": [
    {
      "url": "http://api.egolia.localhost"
    }
  ],
...
  "paths": {
    "/course/certificates/{certificateId}": {
      "get": {
        "security": [
          {
            "course_Oauth2": [
              "openid",
              "entitlements"
            ]
          },
          {
            "course_OIDC": [
              "openid",
              "entitlements"
            ]
          }
        ],
        "tags": [
          "Certificate"
        ],
        "summary": "Get certificate by id",
        "operationId": "getCertificateById",
        "description": "- authenticated:\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/course_certificateIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Certificate detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/course_Certificate"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/course_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/course_UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/course_ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/course_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/course_InternalServerError"
          }
        }
      }
    },
    "/course/certificates/me": {
      "get": {
        "security": [
          {
            "course_Oauth2": [
              "openid",
              "entitlements"
            ]
          },
          {
            "course_OIDC": [
              "openid",
              "entitlements"
            ]
          }
        ],
        "tags": [
          "Certificate"
        ],
        "summary": "Get my certificates",
        "operationId": "getMyCertificates",
        "description": "- authenticated\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/course_pageQuery"
          },
          {
            "$ref": "#/components/parameters/course_limitQuery"
          },
          {
            "$ref": "#/components/parameters/course_orderQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "My certificates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/course_Certificate"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/course_Pagination"
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/course_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/course_UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/course_InternalServerError"
          }
        }
      }
    },
    "/course/certificates": {
      "post": {
        "security": [
          {
            "course_Oauth2": [
              "openid",
              "entitlements"
            ]
          },
          {
            "course_OIDC": [
              "openid",
              "entitlements"
            ]
          }
        ],
        "tags": [
          "Certificate"
        ],
        "summary": "Create certificate",
        "operationId": "createCertificate",
        "description": "- authenticated:\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "courseId",
                  "userId"
                ],
                "properties": {
                  "courseId": {
                    "$ref": "#/components/schemas/course_id"
                  },
                  "userId": {
                    "$ref": "#/components/schemas/course_properties-id"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Certificate created",
            "headers": {
              "Content-Location": {
                "schema": {
                  "type": "string",
                  "format": "uri",
                  "example": "http://api.egolia.localhost/course/certificates/f56b56cb-4dc9-4f84-b9d0-f20f73de2ffd"
                },
                "required": true
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/course_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/course_UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/course_ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/course_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/course_InternalServerError"
          }
        }
      }
    },
    "/course/my-courses": {
      "get": {
        "security": [
          {
            "course_Oauth2": [
              "openid",
              "entitlements"
            ]
          },
          {
            "course_OIDC": [
              "openid",
              "entitlements"
            ]
          }
        ],
        "tags": [
          "Course"
        ],
        "summary": "Get my courses",
        "operationId": "getMyCourses",
        "description": "- display courses of the authenticated instructor .\n- authenticated:\n  - instructor\n  - admin\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/course_pageQuery"
          },
          {
            "$ref": "#/components/parameters/course_limitQuery"
          },
          {
            "$ref": "#/components/parameters/course_orderQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "My courses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/course_Course"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/course_Pagination"
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/course_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/course_UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/course_ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/course_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/course_InternalServerError"
          }
        }
      }
    },
    "/course/courses-published": {
      "get": {
        "security": [],
        "tags": [
          "Course"
        ],
        "summary": "Get published courses",
        "operationId": "getPublishedCourses",
        "description": "- public\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/course_instructorIdQuery"
          },
          {
            "$ref": "#/components/parameters/course_pageQuery"
          },
          {
            "$ref": "#/components/parameters/course_limitQuery"
          },
          {
            "$ref": "#/components/parameters/course_orderQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Published courses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/course_Course"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/course_Pagination"
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/course_BadRequestError"
          },
          "500": {
            "$ref": "#/components/responses/course_InternalServerError"
          }
        }
      }
    },
    "/course/courses-system": {
      "get": {
        "security": [
          {
            "course_Oauth2": [
              "openid",
              "entitlements"
            ]
          },
          {
            "course_OIDC": [
              "openid",
              "entitlements"
            ]
          }
        ],
        "tags": [
          "Course"
        ],
        "summary": "Get system courses",
        "operationId": "getSystemCourses",
        "description": "- authenticated:\n  - instructor\n  - admin\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/course_pageQuery"
          },
          {
            "$ref": "#/components/parameters/course_limitQuery"
          },
          {
            "$ref": "#/components/parameters/course_orderQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Courses in system",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/course_Course"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/course_Pagination"
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/course_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/course_UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/course_ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/course_InternalServerError"
          }
        }
      }
    },
    "/course/reviews/{reviewId}": {
      "put": {
        "security": [
          {
            "course_Oauth2": [
              "openid",
              "entitlements"
            ]
          },
          {
            "course_OIDC": [
              "openid",
              "entitlements"
            ]
          }
        ],
        "tags": [
          "Course"
        ],
        "summary": "Update review",
        "operationId": "updateReview",
        "description": "- authenticated:\n  - review owner\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/course_reviewIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "rating",
                  "comment"
                ],
                "properties": {
                  "rating": {
                    "type": "integer",
                    "format": "int32",
                    "minimum": 1,
                    "maximum": 5,
                    "example": 5
                  },
                  "comment": {
                    "type": "string",
                    "maxLength": 2000,
                    "example": "Great course with practical examples."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Review updated"
          },
          "400": {
            "$ref": "#/components/responses/course_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/course_UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/course_ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/course_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/course_InternalServerError"
          }
        }
      },
      "delete": {
        "security": [
          {
            "course_Oauth2": [
              "openid",
              "entitlements"
            ]
          },
          {
            "course_OIDC": [
              "openid",
              "entitlements"
            ]
          }
        ],
        "tags": [
          "Course"
        ],
        "summary": "Delete review",
        "operationId": "deleteReview",
        "description": "- authenticated:\n  - review owner\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/course_reviewIdPath"
          }
        ],
        "responses": {
          "204": {
            "description": "Review deleted"
          },
          "400": {
            "$ref": "#/components/responses/course_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/course_UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/course_ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/course_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/course_InternalServerError"
          }
        }
      }
    },
    "/course/courses/me/bookmarked": {
      "get": {
        "security": [
          {
            "course_Oauth2": [
              "openid",
              "entitlements"
            ]
          },
          {
            "course_OIDC": [
              "openid",
              "entitlements"
            ]
          }
        ],
        "tags": [
          "Course"
        ],
        "summary": "Get my bookmarked courses",
        "operationId": "getMyBookmarkedCourses",
        "description": "- authenticated:\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/course_pageQuery"
          },
          {
            "$ref": "#/components/parameters/course_limitQuery"
          },
          {
            "$ref": "#/components/parameters/course_orderQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "My bookmarked courses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/course_Course"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/course_Pagination"
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/course_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/course_UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/course_InternalServerError"
          }
        }
      }
    },
...
    "/course/courses/{courseId}/approve": {
      "post": {
        "security": [
          {
            "course_Oauth2": [
              "openid",
              "entitlements"
            ]
          },
          {
            "course_OIDC": [
              "openid",
              "entitlements"
            ]
          }
        ],
        "tags": [
          "Course"
        ],
        "summary": "Approve course",
        "operationId": "approveCourse",
        "description": "- authenticated:\n  - admin\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/course_courseIdPath"
          }
        ],
        "responses": {
          "204": {
            "description": "Course approved"
          },
          "400": {
            "$ref": "#/components/responses/course_BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/course_UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/course_ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/course_NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/course_InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "course_OIDC": {
        "type": "openIdConnect",
        "openIdConnectUrl": "http://authentik.egolia.localhost/application/o/api-web/.well-known/openid-configuration",
        "description": "OpenID Connect Security Scheme",
        "x-defaultClientId": "api-web",
        "x-scalar-client-id": "api-web"
      },
      "course_Oauth2": {
        "type": "oauth2",
        "description": "OAuth2 Security Scheme",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://authentik.notopia.localhost/application/o/authorize/",
            "tokenUrl": "http://authentik.notopia.localhost/application/o/token/",
            "scopes": {
              "openid": "OpenID Connect scope",
              "email": "Email scope",
              "profile": "Profile scope",
              "entitlements": "Entitlements scope"
            },
            "x-defaultClientId": "api-web",
            "x-scalar-client-id": "api-web",
            "x-usePkce": "SHA-256",
            "x-scalar-credentials-location": "header"
          },
          "password": {
            "tokenUrl": "http://authentik.notopia.localhost/application/o/token/",
            "scopes": {
              "openid": "OpenID Connect scope",
              "email": "Email scope",
              "profile": "Profile scope",
              "entitlements": "Entitlements scope"
            },
            "x-default-scopes": [
              "openid",
              "profile",
              "email",
              "entitlements"
            ],
            "x-defaultClientId": "api-web",
...
  },
...
}